execute container exec method
